requirements.txt
requests
lxml
pyserial